consumer/requirements.txt
confluent-kafka==2.3.0
pymongo==4.6.0
python-dotenv==1.0.0
